Move example utils into pip package
