+++ examples/schelling_false.py
@@ -12,6 +12,7 @@ Taken from: http://forestdb.org/models/schelling-falsebelief.html
 from __future__ import print_function
 
 import argparse
+
 import torch
 
 import pyro
