+++ examples/vae.py
@@ -1,4 +1,5 @@
 import argparse
+
 import numpy as np
 import torch
 import torch.nn as nn
@@ -6,12 +7,11 @@ import visdom
 
 import pyro
 import pyro.distributions as dist
+from pyro.contrib.examples.mnist_cached import MNISTCached as MNIST
+from pyro.contrib.examples.mnist_cached import setup_data_loaders
+from pyro.contrib.examples.vae_plots import mnist_test_tsne, plot_llk, plot_vae_samples
 from pyro.infer import SVI
 from pyro.optim import Adam
-from utils.vae_plots import plot_llk, mnist_test_tsne, plot_vae_samples
-from utils.mnist_cached import MNISTCached as MNIST
-from utils.mnist_cached import setup_data_loaders
-
 
 fudge = 1e-7
