+++ pyro/contrib/examples/mnist_cached.py
