+++ pyro/contrib/examples/util.py
