+++ pyro/contrib/examples/vae_plots.py
